jd02_01 fixed retired analysis jd02_02 started jd02_03 НЕ получилось в С открывать кассы по алгоритму, посчитать итоговую сумму выручки и даже посчитать сумму чека (параметр overAll из диспетчера)

+++ src/by/it/toporova/jd02_03/Basket.java
@@ -5,7 +5,7 @@ import java.util.HashMap;
 import java.util.Map;
 
 class Basket {
-    Map<String, Double> goods = new HashMap<>();
+    Map<String, Integer> goods = new HashMap<>();
 
     Basket() {
     }
